Added frontend UI

+++ backend/controller/goalController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../model/goalsModel');
+const User = require('../model/userModel');
 
 
 //  @desc Get goals
@@ -9,7 +10,7 @@ const Goal = require('../model/goalsModel');
 
  const getGoals = asyncHandler(async (req,res) => {
 
-   const goals = await Goal.find();
+   const goals = await Goal.find({user: req.user.id});
 
     res.status(200).json(goals);
  })
@@ -25,7 +26,8 @@ const Goal = require('../model/goalsModel');
    } 
 
    const goal = await Goal.create({
-      text: req.body.text
+      text: req.body.text,
+      user: req.user.id
    })
 
    res.status(200).json(goal);
@@ -42,6 +44,21 @@ const Goal = require('../model/goalsModel');
    if(!goal){
       res.status(400);
       throw new Error('Goal not found');
+   }
+
+   const user = await User.findById(req.user.id);
+
+   // Check for user
+   if(!user){
+      res.status(401)
+      throw new Error('User not found');
+   }
+
+   // Make sure the logged in user matches the goal user
+   if(goal.user.toString() !== user.id)
+   {
+      res.status(401)
+      throw new Error('User not authorized');
    }
 
    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -61,6 +78,21 @@ const Goal = require('../model/goalsModel');
    if(!goal){
       res.status(400);
       throw new Error('Goal not found');
+   }
+
+   const user = await User.findById(req.user.id);
+
+   // Check for user
+   if(!user){
+      res.status(401)
+      throw new Error('User not found');
+   }
+
+   // Make sure the logged in user matches the goal user
+   if(goal.user.toString() !== user.id)
+   {
+      res.status(401)
+      throw new Error('User not authorized');
    }
 
    await goal.deleteOne()
